Moved constants to rules
